flaskblog/requirements.txt
Flask
Flask-sqlalchemy
pillow
Flask-migrate
requests
WtForms
email-validator
Flask-login
Flask-bcrypt
pykakasi